+++ qa/evil-tests/src/test/resources/org/elasticsearch/plugins/cli/complex-plugin-security.policy
