+++ qa/evil-tests/src/test/resources/org/elasticsearch/plugins/cli/simple-plugin-security.policy
